backend/app/requirements.txt
sockets
pycryptodome
sympy
